src/requirements_vgg_Face.txt
absl-py==1.4.0
amqp==5.1.1
asgiref==3.7.2
astroid==2.15.6
astunparse==1.6.3
async-timeout==4.0.2
billiard==3.6.4.0
black==21.12b0
cachetools==5.3.1
celery==5.2.7
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==6.7.0
coverage==7.2.7
crypto==1.4.1
dill==0.3.7
Django==3.0.8
django-createsuperuserwithpassword==2.0.0
django-model-utils==4.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
dlib==19.22.1
docopt==0.6.2
exceptiongroup==1.1.2
face-compare==1.1.4
face-engine==2.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
flake8==6.0.0
Flask==1.1.2
flatbuffers==1.12
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.0.0
Jinja2==3.0.0
joblib==1.0.1
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface @ git+https://github.com/M-Taghizadeh/keras-vggface.git@3d2fb52c316ed1de399c05ac83678f1bb7b8a43b
kombu==5.3.1
lazy-object-proxy==1.9.0
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
Naked==0.1.32
numpy==1.20.2
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
pathspec==0.11.2
Pillow==8.4.0
platformdirs==3.10.0
pluggy==1.2.0
progressbar2==3.53.1
prompt-toolkit==3.0.39
protobuf==3.19.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycryptodome==3.17
pyflakes==3.0.1
PyJWT==2.8.0
pylint==2.16.2
pytest==7.4.0
pytest-cov==4.0.0
pytest-watch==4.2.0
python-dotenv==0.17.0
python-utils==2.5.6
pytz==2023.3
PyYAML==6.0.1
redis==4.5.1
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.24.2
scipy==1.6.3
shellescape==3.8.1
six==1.15.0
sqlparse==0.4.4
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
threadpoolctl==2.1.0
tomli==1.2.3
tomlkit==0.12.1
tqdm==4.65.0
typing-extensions==4.7.1
urllib3==1.26.16
vine==5.0.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.0.3
wrapt==1.15.0
zipp==3.16.2
